refactor(types): extract the ssh api surface from the shared api types
src/shared/types/api.ts is the aggregate every process imports, and it has
grown to the point where any new API method has to argue with the file size
guard before it can argue with a reviewer. Every other API family in it has
already moved out to its own sibling module - ./editor, ./terminal, ./tmux,
./review, ./schedule, ./notifications, ./cliInstaller, ./extensions,
./visualization - and SSH is the last one still inline.

This moves the SSH block verbatim into src/shared/types/ssh.ts and leaves
api.ts with the single `import type { SshAPI }` that `ElectronAPI` needs. The
moved lines are unchanged; only the module header and the re-export are new.
`@shared/types` re-exports the new module, so every consumer that already
imported through the barrel is untouched. The one deep importer,
ConfigManager, now goes through the barrel like its neighbours.

api.ts: 1158 -> 1073 lines.
